remove integration helper and policy separation
