+++ tests/integration/conftest.py
@@ -13,7 +13,8 @@ from helpers import (
     METADATA,
     NGINX_NAME,
     POSTGRES_NAME,
-    perform_ranger_integrations,
+    TRINO_NAME,
+    TRINO_SERVICE,
 )
 from pytest_operator.plugin import OpsTest
 
@@ -31,8 +32,14 @@ async def deploy(ops_test: OpsTest):
         ]
     }
     await ops_test.model.deploy(POSTGRES_NAME, channel="14", trust=True)
-    ranger_config = {"user-group-configuration": GROUP_MANAGEMENT}
+    await ops_test.model.wait_for_idle(
+        apps=[POSTGRES_NAME],
+        status="active",
+        raise_on_blocked=False,
+        timeout=1000,
+    )
 
+    ranger_config = {"user-group-configuration": GROUP_MANAGEMENT}
     await ops_test.model.deploy(
         charm,
         resources=resources,
@@ -40,31 +47,65 @@ async def deploy(ops_test: OpsTest):
         num_units=1,
         config=ranger_config,
     )
+
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME],
+        status="blocked",
+        raise_on_blocked=False,
+        timeout=1000,
+    )
+
+    await ops_test.model.integrate(APP_NAME, POSTGRES_NAME)
+
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME, POSTGRES_NAME],
+        status="active",
+        raise_on_blocked=False,
+        timeout=1500,
+    )
+
+    trino_config = {
+        "charm-function": "all",
+        "ranger-service-name": TRINO_SERVICE,
+    }
+
+    await ops_test.model.deploy(
+        TRINO_NAME,
+        channel="beta",
+        config=trino_config,
+        trust=True,
+    )
+
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME, TRINO_NAME],
+        status="active",
+        raise_on_blocked=False,
+        timeout=1500,
+    )
+    await ops_test.model.integrate(APP_NAME, TRINO_NAME)
+
+    await ops_test.model.wait_for_idle(
+        apps=[APP_NAME, TRINO_NAME],
+        status="active",
+        raise_on_blocked=False,
+        timeout=1500,
+    )
     await ops_test.model.deploy(NGINX_NAME, trust=True)
+    await ops_test.model.wait_for_idle(
+        apps=[NGINX_NAME],
+        status="active",
+        raise_on_blocked=False,
+        timeout=1500,
+    )
 
-    async with ops_test.fast_forward():
-        await ops_test.model.wait_for_idle(
-            apps=[POSTGRES_NAME],
-            status="active",
-            raise_on_blocked=False,
-            timeout=1000,
-        )
-
-        await ops_test.model.wait_for_idle(
-            apps=[APP_NAME],
-            status="blocked",
-            raise_on_blocked=False,
-            timeout=1000,
-        )
-        await perform_ranger_integrations(ops_test, APP_NAME)
-
-        await ops_test.model.wait_for_idle(
-            apps=[NGINX_NAME, APP_NAME],
-            status="active",
-            raise_on_blocked=False,
-            timeout=1000,
-        )
-        assert (
-            ops_test.model.applications[APP_NAME].units[0].workload_status
-            == "active"
-        )
+    await ops_test.model.integrate(APP_NAME, NGINX_NAME)
+    await ops_test.model.wait_for_idle(
+        apps=[NGINX_NAME, APP_NAME],
+        status="active",
+        raise_on_blocked=False,
+        timeout=1000,
+    )
+    assert (
+        ops_test.model.applications[APP_NAME].units[0].workload_status
+        == "active"
+    )
